splitted utils on common and render
