+++ src/view/film-card.js
@@ -1,5 +1,5 @@
 import AbstractComponent from './abstract-component.js';
-import {formatDuration, getFileName} from '../helpers/utils';
+import {formatDuration, getFileName} from '../helpers/common';
 
 
 const setControlClass = (control) => control ? `film-card__controls-item--active` : ``;
