+++ src/view/film-details.js
@@ -1,5 +1,5 @@
 import AbstractComponent from './abstract-component.js';
-import {formatDuration, getFileName} from '../helpers/utils';
+import {formatDuration, getFileName} from '../helpers/common';
 
 const createFilmDetails = (filmCard) => {
   const {
